+++ src/components/not_found/index.tsx
@@ -2,7 +2,10 @@ import React from 'react';
 import classnames from 'classnames';
 import useTranslation from 'next-translate/useTranslation';
 import { Typography } from '@material-ui/core';
-import { Face } from '@material-ui/icons';
+import { useRecoilValue } from 'recoil';
+import { readTheme } from '@recoil/settings/selectors';
+import NotFoundLight from '@assets/not-found-light.svg';
+import NotFoundDark from '@assets/not-found-dark.svg';
 import { useStyles } from './styles';
 
 const NotFound: React.FC<{
@@ -10,10 +13,16 @@ const NotFound: React.FC<{
 }> = ({ className }) => {
   const classes = useStyles();
   const { t } = useTranslation('common');
+  const theme = useRecoilValue(readTheme);
+
   return (
     <div className={classnames(className, classes.root)}>
       <div className={classes.content}>
-        <Face />
+        {theme === 'light' ? (
+          <NotFoundLight />
+        ) : (
+          <NotFoundDark />
+        )}
         <Typography variant="body1">
           {t('notFound')}
         </Typography>