+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Unreleased
+
+## Changes
+- Updated not found and 404 logo ([\#792](https://github.com/forbole/big-dipper-2.0-cosmos/issues/792))
+- Fixed params % display ([\#795](https://github.com/forbole/big-dipper-2.0-cosmos/issues/795))
+## Fixes
+- Fixed commission being displayed for non validator accounts ([\#787](https://github.com/forbole/big-dipper-2.0-cosmos/issues/787))
+
+# base-v2.0.1 - 2021-03-17
+## Bug fixes
+- Fixed online voting power total ([\#800](https://github.com/forbole/big-dipper-2.0-cosmos/issues/800))
+
 # base-v2.0.0 - 2021-03-10
 
 ## Changes
@@ -7,6 +19,9 @@
 - Improved initial loading speed by asyncing top level functions ([\#773](https://github.com/forbole/big-dipper-2.0-cosmos/issues/773))
 ## Bug fixes
 - Fixed Apr error if bonded tokens is 0 ([\#758](https://github.com/forbole/big-dipper-2.0-cosmos/issues/758))
+
+## Breaking
+- [Bdjuno](https://github.com/forbole/bdjuno) must be on `v2.0.0`
 
 # base-v2.0.0-rc3 - 2021-02-25
 
@@ -19,8 +34,7 @@
 - Change env `NEXT_PUBLIC_CHAIN_STATUS` to `NEXT_PUBLIC_CHAIN_TYPE` or don't. It's backwards compatible
 
 ## Breaking
-
-- [Bdjuno](https://github.com/forbole/bdjuno) must be on `v1.1.0+`
+- [Bdjuno](https://github.com/forbole/bdjuno) must be on `v2.0.0`
 
 # base-v2.0.0-rc2 - 2021-02-24